Prevent interleaved execution of classes with method dependencies
TestNG implements dependencies by running all "free" methods
first, then running all methods that just became "free", a.s.o. This
results in having multiple test classes active simultaneously. This is
undesirable when the tests share resources such as a database and the
classes are independent and intended to run sequentially.

This fix bypasses other TestNG ordering features and runs methods in
order of dependencies. As a result, classes are executed sequentially
as long as dependencies do not cross class boundaries.

This breaks the @Test.priority annotation and the @Listeners annotation
with an IMethodInterceptor and therefore is not appropriate for
submission upstream.